# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Find Your Car",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/fyc/UserAudit": {
            "get": {
                "description": "Get a list of UserAudit or a specific UserAudit by ID with optional extra data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Audit"
                ],
                "summary": "Get UserAudit or specific UserAudit by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserAudit ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of UserAudit or a single UserAudit",
                        "schema": {
                            "$ref": "#/definitions/pkg.UserAudit"
                        }
                    },
                    "400": {
                        "description": "Bad request: Invalid UserAudit ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No UserAudit found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing UserAudit by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Audit"
                ],
                "summary": "Update a UserAudit by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserAudit ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated UserAudit data",
                        "name": "UserAudit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.UserAudit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserAudit updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or ID mismatch",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "UserAudit not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to update UserAudit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new UserAudit to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Audit"
                ],
                "summary": "Add a new UserAudit",
                "parameters": [
                    {
                        "description": "UserAudit data",
                        "name": "UserAudit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.UserAudit"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "UserAudit created successfully",
                        "schema": {
                            "$ref": "#/definitions/pkg.UserAudit"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to create a new UserAudit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete a UserAudit by setting the is_deleted flag to true",
                "tags": [
                    "User Audit"
                ],
                "summary": "Soft delete a UserAudit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserAudit ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserAudit deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid UserAudit ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to delete UserAudit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.UserAudit": {
            "type": "object",
            "required": [
                "action_date",
                "module",
                "new_value",
                "old_value",
                "user_id"
            ],
            "properties": {
                "action_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "module": {
                    "type": "string"
                },
                "new_value": {
                    "type": "object"
                },
                "old_value": {
                    "type": "object"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}